Add base impls

+++ src/main/java/com/kelaidisc/domain/Course.java
@@ -2,8 +2,8 @@ package com.kelaidisc.domain;
 
 public class Course {
 
-    private int id;
-    private String name;
-    private String description;
+  private Long id;
+  private String name;
+  private String description;
 
 }
